ID-sync
platforms other than bsd should check and see if
they can make the CONTENDED macro in the lock_h directory
work. We are now using contention on the lock as an
additional indication that there is a pending reference
to an inp...

If your locking system can tell you someone is waiting
on the lock you should put that test in instead of the (0)
I added.