This adds in support for the NDATA chunk. It also takes and make some of the testers and such compile (the stream reset changes are commented out).
Note to Andreas and others, if you use this
please *do not* enable the NDATA chunk. If you
do and more than one re-assembled message is
sent it will break. A new branch will soon
appear that takes on the task of re-writing the
reassembly process.. which we are in desparate
need of doing.